Cleanup rebase process
