+++ gdal/src/main/scala/geotrellis/contrib/vlm/gdal/GDALDataType.scala
